+++ src/Framework.Configurator/configurator-ui/src/app/rights/components/principals/components/grant-rights-dialog/grant-rights-dialog.component.ts
@@ -8,7 +8,8 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 import { SearchFieldComponent } from './components/search-header/search-header.component';
 import { ContextFilterPipe } from './context-filter.pipe';
-import { ContextsApiService, PrincipalApiService } from 'src/app/shared/api.services';
+import { PrincipalApiService } from 'src/app/shared/api-services/principal.api.service';
+import { ContextsApiService } from 'src/app/shared/api-services/context.api.serivce';
 import { RightsFilterPipe } from './rights-filter.pipe';
 import { DestroyService } from 'src/app/shared/destroy.service';
 import { MatMenuModule } from '@angular/material/menu';
