+++ src/Framework.Configurator/configurator-ui/src/app/rights/components/principals/components/permission-edit-dialog/add-entity-control/add-entity-control.component.ts
@@ -6,7 +6,7 @@ import { MatIconModule } from '@angular/material/icon';
 import { Observable, of, filter, switchMap, tap, startWith, debounceTime, Subject, BehaviorSubject } from 'rxjs';
 import { IEntity } from '../../view-principal-dialog/view-principal-dialog.component';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { ContextsApiService } from 'src/app/shared/api.services';
+import { ContextsApiService } from 'src/app/shared/api-services/context.api.serivce';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
